@@ -5,42 +5,32 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "report": "vue-cli-service build --report"
   },
   "dependencies": {
-    "register-service-worker": "^1.0.0",
-    "vue": "^2.5.16",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1",
-    "axios": "^0.18.0",
-    "firebase": "^5.1.0",
-    "firebaseui": "^3.1.1",
-    "vue-analytics": "^5.12.2",
-    "vue-infinite-loading": "^2.3.1",
+    "axios": "^0.18.1",
+    "firebase": "^5.11.1",
+    "firebaseui": "^3.6.1",
+    "register-service-worker": "^1.6.2",
+    "vue": "^2.6.10",
+    "vue-analytics": "^5.16.4",
+    "vue-infinite-loading": "^2.4.4",
     "vue-jsonp": "^0.1.8",
     "vue-lazyload": "^1.2.3",
+    "vue-router": "^3.0.6",
     "vuefire": "^1.4.5",
-    "vuetify": "1.1.8"
+    "vuetify": "1.5.14",
+    "vuetify-loader": "1.2.1",
+    "vuex": "^3.1.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0-rc.8",
-    "@vue/cli-plugin-eslint": "^3.0.0-rc.8",
-    "@vue/cli-plugin-pwa": "^3.0.0-rc.8",
-    "@vue/cli-service": "^3.0.0-rc.8",
-    "lint-staged": "^7.2.0",
-    "vue-template-compiler": "^2.5.16"
-  },
-  "gitHooks": {
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.js": [
-      "vue-cli-service lint",
-      "git add"
-    ],
-    "*.vue": [
-      "vue-cli-service lint",
-      "git add"
-    ]
+    "@vue/cli-plugin-babel": "^3.7.0",
+    "@vue/cli-plugin-eslint": "^3.7.0",
+    "@vue/cli-plugin-pwa": "^3.7.0",
+    "@vue/cli-service": "^3.7.0",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.2",
+    "vue-template-compiler": "^2.6.10"
   }
 }
