@@ -4,12 +4,21 @@ module.exports = {
     node: true
   },
   'extends': [
-    'plugin:vue/essential',
+    'plugin:vue/strongly-recommended',
     'eslint:recommended'
   ],
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
+    'no-console': 'off',
+    'no-debugger': 'off',
+    'vue/attribute-hyphenation': 'off',
+    'vue/attributes-order': 'error',
+    'vue/script-indent': 'error',
+    'vue/html-closing-bracket-newline': [
+      'error',
+      {
+        multiline: 'always'
+      }
+    ]
   },
   parserOptions: {
     parser: 'babel-eslint'
