@@ -1,8 +1,40 @@
 <template>
-  <v-app v-bind="theme" :class="blackClass">
+  <v-app 
+    v-bind="theme"
+    :class="blackClass"
+  >
+    <context-menu 
+      ref="con" 
+      :key="uid"
+    />
+    <snackbar/>
+    <v-snackbar
+      v-model="snackbar"
+      :timeout="0"
+      top
+      auto-height
+    >
+      A newer version of dreamcloud has been downloaded
+      <v-btn
+        dark
+        flat
+        @click="() => swReg.waiting.postMessage('skipWaiting')"
+      >
+        use now
+      </v-btn>
+      <v-btn
+        dark
+        flat
+        @click="snackbar = false"
+      >
+        dismiss
+      </v-btn>
+    </v-snackbar>
+    
     <!-- left drawer -->
     <v-navigation-drawer
       id="sideleft"
+      v-model="drawLeft"
       app
       clipped
       disable-route-watcher
@@ -10,39 +42,54 @@
       persistent
       ripple
       width="240"
-      v-model="drawLeft">
-      <sidebar @closeLeft="closeLeft"></sidebar>
+    >
+      <sidebar @closeLeft="closeLeft"/>
     </v-navigation-drawer>
 
     <!-- header -->
-    <v-toolbar app fixed clipped-left clipped-right dense id="navbar">
-      <v-progress-linear
+    <v-toolbar 
+      id="navbar" 
+      app 
+      dense 
+      fixed 
+      clipped-left 
+      clipped-right
+    >
+      <!-- <v-progress-linear
         id="loader"
-        color="primary"
-        background-color="#"
-        height="3"
         :active="loadActive"
         :value="loadValue"
-      ></v-progress-linear>
+        color="primary"
+        background-color="#"
+        height="2"
+      /> -->
 
+      <!-- debug button -->
+      <!-- <v-toolbar-side-icon @click.stop="$store.dispatch('snack', { b: true, c:'green', s:'wat' })"/> -->
       <!-- toggle left draw button -->
-      <v-toolbar-side-icon @click.stop="leftTog"></v-toolbar-side-icon>
+      <v-toolbar-side-icon @click.stop="leftTog"/>
       
       <!-- title -->
-      <v-toolbar-title class="hidden-sm-and-down fwl title" style="width: 170px"  >
-        <router-link :class="textClass" :to="{name:'historyRecommended', params: {user: $DCFB.UID}}">
-          dreamcloud 
+      <v-toolbar-title 
+        class="hidden-sm-and-down fwl title" 
+        style="width: 121px"
+      >
+        <router-link 
+          :class="textClass" 
+          :to="{name:'about', params: {user: uid}}"
+        >
+          dreamcloud
         </router-link>
       </v-toolbar-title>
 
-      <!-- <v-spacer></v-spacer> -->
+      <search/>
 
-      <!-- searchbar -->
-      <search></search>
-
-      <v-spacer></v-spacer>
+      <v-spacer/>
       <!-- toggle stage button -->
-      <v-toolbar-side-icon v-if="!bMobi" @click.stop="$store.commit('toggleStage')"><v-icon>music_video</v-icon></v-toolbar-side-icon>
+      <!-- <v-toolbar-side-icon 
+        v-if="!bMobi" 
+        @click.stop="$store.commit('toggleStage')"
+      ><v-icon>music_video</v-icon></v-toolbar-side-icon> -->
 
       <!-- toggle right draw button -->
       <v-toolbar-side-icon @click.stop="rightTog"><v-icon>playlist_play</v-icon></v-toolbar-side-icon>
@@ -50,24 +97,34 @@
 
     <!-- right drawer -->
     <v-navigation-drawer
+      id="right-draw"
+      v-model="drawRight"
       app
       clipped
       disable-route-watcher
       disable-resize-watcher
       persistent
       right
-      v-model="drawRight"
       width="260"
-      id="right-draw">
-      <current-playlist></current-playlist>
+    >
+      <current-playlist @conmen="con"/>
     </v-navigation-drawer>
 
     <!-- Stage -->
-    <v-content class="text-xs-center" v-show="bShowStage">
-      <v-container fluid fill-height class="pa-0">
+    <v-content 
+      v-show="bShowStage" 
+      class="text-xs-center"
+    >
+      <v-container 
+        fluid 
+        fill-height 
+        class="pa-0"
+      >
         <v-layout justify-center>
           <v-flex xs12>
-            <stage></stage>
+            <stage 
+              @conmen="con"
+            />
           </v-flex>
         </v-layout>
       </v-container>
@@ -75,156 +132,228 @@
 
     <!-- Router view -->
     <v-content class="text-xs-center">
-      <v-container fluid fill-height id="main-cont" class="pa-0 p1t-4 pb1-4">
+      <v-container 
+        id="main-cont" 
+        fluid 
+        fill-height 
+        class="pa-0 ma-0"
+      >
         <v-layout justify-center>
-          <!-- <transition name="  fade" mode="out-in" :key="$route.fullPath"> -->
-            <!-- <keep-alive> -->
-              <router-view></router-view>
-            <!-- </keep-alive> -->
+          <!-- <transition 
+            :key="$route.fullPath" 
+            name="  fade" 
+            mode="out-in"> -->
+          <keep-alive 
+            :max="2"
+          >
+            <router-view @conmen="con"/>
+          </keep-alive>
           <!-- </transition> -->
         </v-layout>
       </v-container>
     </v-content>
 
     <!-- Footer -->
-    <v-footer app fixed id="foot" :style="footStyle">
-      <mobileFooter v-show="bMobi"></mobileFooter>
-      <dc-youtube v-show="(ytUseVideo && isYT) && (!bMobi || currentActive)"></dc-youtube>
-      <dc-audio v-show="(!ytUseVideo || !isYT) && (!bMobi || currentActive)"></dc-audio>
-      <scroll-to-top v-if="!bMobi"></scroll-to-top>
-    </v-footer>
+    <v-footer 
+      id="foot" 
+      :style="footStyle" 
+      app 
+      fixed
+    >
+      <mobileFooter v-if="bMobi"/>
+      <!-- && (bMobi || currentActive) -->
+      <dc-youtube v-show="(ytUseVideo && isYT)"/>
+      <dc-audio v-show="(!ytUseVideo || !isYT)"/>
+      <!-- <v-flex xs12 xl12> -->
+      <!-- <v-btn block large color="primary ">Get started</v-btn> -->
+      <!-- </v-flex> -->
 
+    </v-footer>
     <!-- dc keyboard shortcuts -->
-    <hks></hks>
+    <hks/>
   </v-app>
 </template>
 
 <script>
-  import hks from './components/misc/hks'
-  import search from './components/header/search'
-  import dcAudio from './components/footer/dc-audio'
-  import dcAudio2 from './components/footer/dc-audio-flex-2.vue'
-  import dcYoutube from './components/footer/dc-youtube'
-  import currentPlaylist from './components/sidebar-right/current-playlist'
-  import sidebar from './components/sidebar-left/sidebar'
-  import stage from '@/components/stage/stage'
-  import scrollToTop from '@/components/footer/scroll-to-top.vue'
-  import mobileFooter from '@/components/footer/mobileFooter'
-  import { mapGetters } from 'vuex'
+import hks from './components/misc/hks'
+import search from './components/header/search'
+import snackbar from './components/header/snackbar'
+import dcAudio from './components/footer/dc-audio'
+import dcYoutube from './components/footer/dc-youtube'
+import currentPlaylist from './components/sidebar-right/current-playlist'
+import sidebar from './components/sidebar-left/sidebar'
+import stage from '@/components/stage/stage'
+import mobileFooter from '@/components/footer/mobileFooter'
 
-  export default {
-    name: 'app',
-    components: {
-      'hks': hks,
-      'search': search,
-      'dc-audio2': dcAudio2,
-      'dc-audio': dcAudio,
-      'dc-youtube': dcYoutube,
-      'sidebar': sidebar,
-      'current-playlist': currentPlaylist,
-      'stage': stage,
-      'mobileFooter': mobileFooter,
-      'scroll-to-top': scrollToTop
+import { mapGetters } from 'vuex'
+
+export default {
+  name: 'App',
+  data () {
+    return {
+      snackbar: false,
+      swReg: '',
+      refreshing: false
+    }
+  },
+  components: {
+    'hks': hks,
+    'snackbar': snackbar,
+    'search': search,
+    'dc-audio': dcAudio,
+    'dc-youtube': dcYoutube,
+    'sidebar': sidebar,
+    'current-playlist': currentPlaylist,
+    'stage': stage,
+    'mobileFooter': mobileFooter
+  },
+  methods: {
+    con (a) {
+      this.$refs.con.show(a[0], a[1])
     },
-    methods: {
-      leftTog () {
-        this.$store.commit('drawLeftTog')
-      },
-      rightTog () {
-        this.$store.commit('drawRightTog')
-      },
-      closeLeft () {
-        if (this.$UTILS.isMobile) {
-          this.drawerLeft = false
-        }
+    leftTog () {
+      this.$store.commit('drawLeftTog')
+    },
+    rightTog () {
+      this.$store.commit('drawRightTog')
+    },
+    closeLeft () {
+      if (this.$UTILS.isMobile) {
+        this.$store.commit('drawLeft', false)
+        // console.log('toggle left')
+        this.drawLeft = false
       }
     },
-    computed: {
-      footStyle () {
-        return {
-          height: (this.currentActive && this.bMobi ? '128px' : '56px') + ' !important'
-        }
-      },
-      bMobi () {
-        return this.$vuetify.breakpoint.name === 'xs' || this.$vuetify.breakpoint.name === 'sm'
-      },
-      textClass () {
-        return (this.$store.getters.nightMode ? 'white' : 'black') + '--text noDeco pointer hidden-sm-and-down'
-      },
-      currentActive () {
-        return this.$route.name === 'stage'
-      },
-      drawRight: {
-        get () {
-          return this.$store.getters.drawRight
-        },
-        set (value) {
-          this.$store.commit('drawRight', value)
-        }
-      },
-      drawLeft: {
-        get () {
-          return this.$store.getters.drawLeft
-        },
-        set (value) {
-          this.$store.commit('drawLeft', value)
-        }
-      },
-      ...mapGetters({
-        blackClass: 'blackClass',
-        isYT: 'isYT',
-        ytUseVideo: 'ytUseVideo',
-        bShowStage: 'bShowStage',
-        theme: 'theme',
-        loadValue: 'loadValue',
-        loadActive: 'loadActive',
-        ytFullScreen: 'ytFullScreen'
-      })
+    listenForWaitingServiceWorker (reg, callback) {
+      function awaitStateChange() {
+        reg.installing.addEventListener('statechange', function() {
+          if (this.state === 'installed') callback(reg);
+        });
+      }
+      if (!reg) return;
+      if (reg.waiting) return callback(reg);
+      if (reg.installing) awaitStateChange();
+      reg.addEventListener('updatefound', awaitStateChange);
+    }
+  },
+  computed: {
+    ...mapGetters({
+      uid: 'uid',
+      blackClass: 'blackClass',
+      isYT: 'isYT',
+      ytUseVideo: 'ytUseVideo',
+      bShowStage: 'bShowStage',
+      theme: 'theme',
+      loadValue: 'loadValue',
+      loadActive: 'loadActive',
+      ytFullScreen: 'ytFullScreen'
+    }),
+    footStyle () {
+      return {
+        height: (this.bMobi ? '128px' : '56px') + ' !important'
+      }
+      // return {
+      //   height: (this.currentActive && this.bMobi ? '128px' : '56px') + ' !important'
+      // }
     },
-    beforeCreate () {
-      // Set them
-      this.$vuetify.theme.primary = '#009688'
-      this.$store.commit('ytUseVideo', !this.bMobi)
-      // console.log(this.$vuetify.theme)
-      // On Firebase auth state change
-      this.$DCFB.fb.auth().onAuthStateChanged((user) => {
-        // console.log(this.$DCFB.fb.auth().currentUser)
-        // Toggle auth state to false
-        this.$store.commit('authChange', false)
-        this.$nextTick(() => {
-          // Set auth state
-          this.$store.commit('authChange', !!user)
-          // If logged in
-          if (!user) {
-            // Sign in anonymously
-            this.$DCFB.fb.auth().signInAnonymously()
-          } else {
-            // Update store
-            this.$store.commit('setUser', user)
-            if (user.isAnonymous) {
-              this.$router.push({name: 'about'})
-            } else {
-              if (!this.$route.name) {
-                this.$router.push({name: 'historyRecommended'})
-              }
-            }
-            // Initialise DCFB plugin
-            this.$DCFB.init(user.uid)
-            // Get settings
-            this.$DCFB.settings.once('value', (snapshot) => {
-              if (snapshot.val() !== null) {
-                this.$store.commit('settings', snapshot.val())
-              }
-              // this.listViewSmall = !!snapshot
-            })
-          }
+    bMobi () {
+      return this.$vuetify.breakpoint.name === 'xs' || this.$vuetify.breakpoint.name === 'sm'
+    },
+    textClass () {
+      return (this.$store.getters.nightMode ? 'white' : 'black') + '--text noDeco pointer hidden-sm-and-down'
+    },
+    currentActive () {
+      return this.$route.name === 'stage'
+    },
+    drawRight: {
+      get () {
+        return this.$store.getters.drawRight
+      },
+      set (value) {
+        this.$store.commit('drawRight', value)
+      }
+    },
+    drawLeft: {
+      get () {
+        return this.$store.getters.drawLeft
+      },
+      set (value) {
+        this.$store.commit('drawLeft', value)
+      }
+    }
+  },
+  created () {
+    if ('serviceWorker' in navigator) {
+      navigator.serviceWorker.addEventListener('controllerchange', () => {
+        if (this.refreshing) return
+        this.refreshing = true
+        window.location.reload()
+      }
+      )
+      navigator.serviceWorker.getRegistration().then(e => {
+        this.swReg = e
+        this.listenForWaitingServiceWorker(this.swReg, () => {
+          this.snackbar = true
         })
       })
     }
+    // Set theme
+    this.$vuetify.theme.primary = '#009688'
+    // this.$store.commit('ytUseVideo', !this.bMobi)
+    // On Firebase auth state change
+    this.$DCFB.fb.auth().onAuthStateChanged((user) => {
+      // console.log(this.$DCFB.fb.auth().currentUser)
+      // Toggle auth state to false
+      this.$store.commit('authChange', false)
+      this.$nextTick(() => {
+        // Set auth state
+        this.$store.commit('authChange', !!user) // dirty !! to convert int to bool
+        // If not logged in
+        if (!user) {
+          // Sign in anonymously and wait, WAIT PLS
+          this.$DCFB.fb.auth().signInAnonymously()
+          return
+        } else {
+          // Update store with user uid
+          this.$store.commit('setUser', user)
+        }
+
+        // Initialise DCFB plugin
+        this.$DCFB.init(user.uid)
+
+        // redirect to somewhere meaningful
+        if (user.isAnonymous && !this.$route.name) {
+          this.$router.push({name: 'about'})
+        } else if (!this.$route.name) {
+          this.$router.push({name: 'home', user: user.uid})
+        }
+        // set isMobile in store for router guard (close menus on back)
+        this.$store.commit('isMobile', this.$vuetify.breakpoint.smAndDown)
+
+        // Get settings
+        this.$DCFB.settings.once('value', (snapshot) => {
+          if (snapshot.val() !== null) {
+            this.$store.commit('settings', snapshot.val())
+          }
+          // this.listViewSmall = !!snapshot
+        })
+      })
+    })
   }
+}
 </script>
 
 <style>
+.filter label, .filter input{
+  font-size: 13px;
+}
+  .fl-l{
+    float: left;
+  } 
+
+  .fl-r{
+    float: right;
+  }
   .container .layout:only-child {
       margin-left: 0px !important;
       margin-right: 0px !important;
@@ -267,7 +396,8 @@
   }
   @media only screen and (min-width: 600px){
     #main-cont{
-      /* margin-top: 0px; */
+      padding-bottom: 50px !important;
+      /* margin-bottom: 35px; */
     }
     #foot{
       padding: 0 !important;
@@ -276,15 +406,16 @@
   }
   @media only screen and (max-width: 599px){
     #main-cont{
-      /* margin-top: 16px !important; */
+      padding-bottom: 110px !important;
+      /* margin-top: 106px !important; */
     }
     #foot .v-input {
       margin-top: -10px;
     }
-    .container {
-      padding: 0 !important;
-      margin: 0 !important;
-    }
+    /* .container { */
+      /* padding: 0 !important; */
+      /* margin: 0 !important; */
+    /* } */
     #foot{
       padding: 0 !important;
       /* height: 75px !important; */
@@ -309,7 +440,7 @@
 .hide{
   display: none;
 }
-width */
+
 #right-draw::-webkit-scrollbar {
     width: 1px;
 }
