@@ -1,7 +1,9 @@
-
+import store from '../vuex'
 var firebase = require('firebase/app')
 require('firebase/auth')
 require('firebase/database')
+
+// import store from '../vuex'
 
 // http://2ality.com/2014/09/es6-modules-final.html
 // https://www.firebase.com/docs/web/guide/saving-data.html
@@ -14,7 +16,14 @@ var config = {
   storageBucket: 'dreamcloud-3f276.appspot.com',
   messagingSenderId: '137974869044'
 }
-
+// window.addEventListener('beforeunload', (event) => {
+//   // Cancel the event as stated by the standard.
+//   event.preventDefault();
+//   console.log('wasd')
+//   // Chrome requires returnValue to be set.
+//   event.returnValue = 'wasssd';
+// })
+  
 class DCFB {
   constructor () {
     this.playlists = ''
@@ -32,7 +41,74 @@ class DCFB {
     this.playlists = this.db.ref('users/' + UID + '/PlaylistsData')
     this.playlistsRefs = this.db.ref('users/' + UID + '/PlaylistsNames')
     this.subscriptions = this.db.ref('users/' + UID + '/Subscriptions')
-    this.history = this.db.ref('users/' + UID + '/History')
+    this.fbhistory = this.db.ref('users/' + UID + '/History')
+
+    // if (window.localStorage) {
+
+    // // if not same user clear LS
+    // if(this.getLoc('dcu') !== UID) {
+    //   this.setLoc('playlist', null)
+    //   this.setLoc('playlistrefs', null)
+    //   this.setLoc('subscriptions', null)
+    //   this.setLoc('settings', null)
+    // }
+    // //set current UID
+    // this.setLoc('dcu', UID)
+
+    // // Get settings
+    // this.settings.on('value', (snap) => {
+    //   if (snap.val() !== null) {
+    //     this.setLoc('settings', JSON.stringify(snap.val()))
+    //   } 
+    // })
+
+    // // Get playlistsRefs
+    // this.playlistsRefs.on('value', (snap) => {
+    //   if (snap.val() !== null) {
+    //     this.setLoc('playlistsRefs', JSON.stringify(snap.val()))
+    //   }
+    // })
+
+    // // Get playlists
+    // this.playlists.on('value', (snap) => {
+    //   if (snap.val() !== null) {
+    //     this.setLoc('playlists', JSON.stringify(snap.val()))
+    //   }
+
+    // })
+
+    // // Get subscriptions
+    // this.subscriptions.on('value', (snap) => {
+    //   if (snap.val() !== null) {
+    //     this.setLoc('subscriptions', JSON.stringify(snap.val()))
+    //   }
+    // })
+
+    // // var connectedRef = firebase.database().ref(".info/connected")
+    // // connectedRef.on("value", (snap) => {
+    // if (!window.navigator.onLine) {
+    //   // alert("not connected");
+    //   let a = JSON.parse(this.getLoc('playlists'))
+    //   let b = JSON.parse(this.getLoc('playlistsRefs'))
+    //   let c = JSON.parse(this.getLoc('subscriptions'))
+    //   let d = JSON.parse(this.getLoc('settings'))
+    //   // store.commit('subscriptions', a)
+
+    //   a && this.playlists.update(a)
+    //   b && this.playlistsRefs.update(b)
+    //   c && this.subscriptions.update(c)
+    //   d && this.settings.update(d)
+    // }
+    // // })
+    // }
+  }
+
+  setLoc (a, b) {
+    return window.localStorage.setItem(a, b)
+  }
+
+  getLoc (a) {
+    return window.localStorage.getItem(a)
   }
 
   setting (name) {
@@ -55,13 +131,20 @@ class DCFB {
     this.subscriptions.child(id).remove()
   }
 
-  createNewPlaylist (name, json) {
+  // key allows old reference
+  createNewPlaylist (name, json, key = false) {
     // Create new playlist reference with id.
-    var nameRef = this.playlistsRefs.push({'name': name, name_lower: name.toLowerCase()})
+    let nameRef
+    if (key) {
+      this.playlistsRefs.child(key).set({'name': name, name_lower: name.toLowerCase()})
+      nameRef = key
+    } else {
+      nameRef = this.playlistsRefs.push({'name': name, name_lower: name.toLowerCase()}).ref.key
+    }
     // Using ID + name push new song.
-    this.playlists.child(nameRef.ref.key).set({'name': name, name_lower: name.toLowerCase()})
+    this.playlists.child(nameRef).set({'name': name, name_lower: name.toLowerCase()})
 
-    this.playlistSongAdd(nameRef.ref.key, json)
+    this.playlistSongAdd(nameRef, json)
   }
 
   playlistDelete (playlistId) {
@@ -69,20 +152,22 @@ class DCFB {
     this.playlists.child(playlistId).remove()
   }
   historyPush (json) {
-    // create new reference
-    var songRef = this.history.push()
-    // save song reference in json.key
-    json.key = songRef.key
-    // remove from json['.key'] bc we have to due to vue-fire
-    delete json['.key']
-    // format date to string
-    json.uploaded = json.uploaded.toString()
-    // update fb
-    songRef.set(json)
+    if (!store.getters.isAnon) {
+      // create new reference
+      var songRef = this.fbhistory.push()
+      // save song reference in json.key
+      json.key = songRef.key
+      // remove from json['.key'] bc we have to due to vue-fire
+      delete json['.key']
+      // format date to string
+      json.uploaded = json.uploaded.toString()
+      // update fb
+      songRef.set(json)
+    }
   }
 
   historyClear () {
-    this.history.remove()
+    this.fbhistory.remove()
   }
 
   playlistSongAdd (id, json) {
