@@ -1,18 +1,34 @@
 /* eslint-disable */
-import axios from 'axios'
-import store from '../vuex'
+import axios from "axios";
+import store from "../vuex";
 export default {
-  install (Vue) {
+  install(Vue) {
     var DCPlayer = {
+      // sBase: 'https://dc-mp3-ceewhqzemu.now.sh/api/',
+      sBase: 'http://dreamcloud.mynetgear.com:7000/api/',
       aPlaylist: [],
       iCurrent: 0,
       eAudio: {src:''},
-      error_count: 0,
+      // error_count: 0,
       init () {
         DCPlayer.bindMediaSesssion()
       },
       ePlayer () {
         return document.getElementById('dc-audio')
+      },
+      getAudio (song, hCallback) {
+        // let a = 'https://cors.io/?https://www.saveoffline.com/process/?type=audio&url=' + url
+        var isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+        let a;
+        // IF SAFARI OR NOT YOUTUBE
+        if (isSafari && song.mp32.indexOf("youtube") > -1) {
+          a = DCPlayer.sBase + "v1/stream/?i=" + song.trackID;
+        } else if (song.mp32.indexOf("youtube") > -1) {
+          a = DCPlayer.sBase + "v2/stream/?i=" + song.trackID;
+        } else {
+          a = DCPlayer.sBase + "v3/stream/?i=" + song.mp32;
+        }
+        return hCallback(a)
       },
       next () {
         // Increment unless end of playlist.
@@ -38,31 +54,35 @@ export default {
         DCPlayer.iCurrent = (DCPlayer.iCurrent > 0 ? DCPlayer.iCurrent - 1 : DCPlayer.aPlaylist.length - 1)
         // store.commit('changeIndex', DCPlayer.iCurrent)
         DCPlayer.playIndex(DCPlayer.iCurrent)
+
       },
       playIndex (index) {
+        // triggers loading animation on dc-audio DCPlayer.eAudio.pause()
+        DCPlayer.eAudio.src = ""
+
         DCPlayer.iCurrent = index
-        store.commit('changeIndex', DCPlayer.iCurrent)
+        store.commit('changeIndex', index)
         this.setMediaSession(DCPlayer.aPlaylist[index])
         if (store.getters.ytUseVideo && store.getters.isYT) {
-          // console.log('playing video')
-          DCPlayer.stop()
-          return 
+          return
         } else {
           // console.log('playing audio')
           store.commit('ytStopVideo')
         }
+        store.commit('dcIsLoading', true)
         if(DCPlayer.aPlaylist[index].source == 'SoundCloud')
           return DCPlayer.setAudioSrc(DCPlayer.aPlaylist[index].mp3)
         else
-          return DCPlayer.play_url(DCPlayer.aPlaylist[index].mp32)
+          return DCPlayer.play_url(DCPlayer.aPlaylist[index])
       },
       play_url (sURL) {
         return DCPlayer.getAudio(sURL, DCPlayer.setAudioSrc)
       },
       setAudioSrc (sURL) {
         //append x=error_count if error_count > 0
-        DCPlayer.eAudio.src = sURL + (DCPlayer.error_count ? '&x=' + DCPlayer.error_count : '')
-        // DCPlayer.eAudio.load()
+        DCPlayer.eAudio.src = sURL
+        // console.log(DCPlayer.eAudio)
+        DCPlayer.eAudio.load()
         var play = DCPlayer.play().then(() => {
           // If play then reset error_count
           DCPlayer.error_count = 0
@@ -71,91 +91,110 @@ export default {
         .catch(error => {
           if(error.toString().indexOf('interrupted by a new load request') === -1){
             DCPlayer.error_count++
-            if (DCPlayer.error_count < 6) {
+            if (DCPlayer.error_count < 2) {
               DCPlayer.error()
             } else {
               // console.log('Failed to play 5 times, next song!')
               DCPlayer.error_count = 0
               DCPlayer.next()
             }
+          } else {
+            // console.log('playing audio')
+            store.commit("ytStopVideo");
           }
+          store.commit("dcIsLoading", true);
+          // return DCPlayer.play_url(DCPlayer.aPlaylist[index]);
+          if (DCPlayer.aPlaylist[index].source == "SoundCloud")
+            return DCPlayer.setAudioSrc(DCPlayer.aPlaylist[index].mp3);
+          else return DCPlayer.play_url(DCPlayer.aPlaylist[index]);
         })
-        return play
       },
-      setPlaylist (array) {
-        DCPlayer.aPlaylist = array
+      play_url(song) {
+        return DCPlayer.getAudio(song, DCPlayer.setAudioSrc);
       },
-      setNPlay (array, index) {
-        DCPlayer.setPlaylist(array)
-        DCPlayer.iCurrent = index
-        return DCPlayer.playIndex(index)
+      setAudioSrc(sURL) {
+        //append x=error_count if error_count > 0
+        // DCPlayer.eAudio.src =
+        //   sURL + (DCPlayer.error_count ? "&x=" + DCPlayer.error_count : "");
+        DCPlayer.eAudio.src = sURL;
+        // console.log(DCPlayer.eAudio)
+        DCPlayer.eAudio.load();
+        var play = DCPlayer.play()
+          .then(() => {
+            // If play then reset error_count
+            DCPlayer.error_count = 0;
+          })
+          // Else if the user doesn't choose another track add 1 to error count and try to play again if under 5
+          .catch(error => {
+            if (
+              error.toString().indexOf("interrupted by a new load request") ===
+              -1
+            ) {
+              // console.log('error', error)
+              DCPlayer.error_count++;
+              if (DCPlayer.error_count < 6) {
+                DCPlayer.error();
+              } else {
+                // console.log('Failed to play 5 times, next song!')
+                DCPlayer.error_count = 0;
+                DCPlayer.next();
+              }
+            } else {
+              // console.log('interrupted by a new load request')
+            }
+          });
+        return play;
       },
-      seekBackward () {
+      setPlaylist(array) {
+        DCPlayer.aPlaylist = array;
+      },
+      setNPlay(array, index) {
+        DCPlayer.setPlaylist(array);
+        DCPlayer.iCurrent = index;
+        return DCPlayer.playIndex(index);
+      },
+      seekBackward() {
         if (store.getters.ytUseVideo && store.getters.isYT) {
-          store.getters.ytObject.seekTo(store.getters.ytObject.getCurrentTime() - 10, true)
+          store.getters.ytObject.seekTo(store.getters.ytObject.getCurrentTime() - 10,true);
         } else {
-          DCPlayer.eAudio.currentTime -= 10
+          DCPlayer.eAudio.currentTime -= 10;
         }
       },
-      seekForward () {
+      seekForward() {
         if (store.getters.ytUseVideo && store.getters.isYT) {
-          store.getters.ytObject.seekTo(store.getters.ytObject.getCurrentTime() + 10, true)
+          store.getters.ytObject.seekTo(
+            store.getters.ytObject.getCurrentTime() + 10,
+            true
+          );
         } else {
-          DCPlayer.eAudio.currentTime += 10
+          DCPlayer.eAudio.currentTime += 10;
         }
       },
-      toggleMute () {
+      toggleMute() {
         if (store.getters.ytUseVideo && store.getters.isYT) {
           if (this.$store.getters.ytObject.isMuted()) {
-            this.$store.getters.ytObject.unMute()
+            this.$store.getters.ytObject.unMute();
           } else {
-            this.$store.getters.ytObject.mute()
+            this.$store.getters.ytObject.mute();
           }
         } else {
-          this.eAudio.muted = !this.eAudio.muted
+          this.eAudio.muted = !this.eAudio.muted;
         }
       },
-      togglePlay () {
+      togglePlay() {
         if (store.getters.ytUseVideo && store.getters.isYT) {
           if (store.getters.ytIsPlaying) {
-            store.getters.ytObject.pauseVideo()
+            store.getters.ytObject.pauseVideo();
           } else {
-            store.getters.ytObject.playVideo()
+            store.getters.ytObject.playVideo();
           }
         } else {
           if (DCPlayer.eAudio.paused) {
-            DCPlayer.eAudio.play()
+            DCPlayer.eAudio.play();
           } else {
-            DCPlayer.eAudio.pause()
+            DCPlayer.eAudio.pause();
           }
         }
-      },
-      getAudio (url, hCallback) {
-        var ax = axios.get('https://www.saveitoffline.com/process/?type=audio&url=' + url)
-        ax.then(function (resp) {
-          if ('data' in resp && 
-          resp.data !== 'Error: no_media_found' && 
-          resp.data !== 'Error: daily_secondary_api_limit_reached' && 
-          resp.data !== 'Error: miss' &&
-          resp.data !== '') {
-
-            for (var i = 0; i < resp.data.urls.length; i++) {
-              if (resp.data.urls[i].label.indexOf('m4a') > -1 || resp.data.urls[i].label.indexOf('audio') > -1) {
-                hCallback(resp.data.urls[i].id)
-                return
-              }
-            }
-            
-            // console.log('fallback', resp.data.urls[Math.max(0,resp.data.urls.length - 2)], resp.data.urls)
-            hCallback(resp.data.urls[Math.max(0,resp.data.urls.length - 3)].id)
-          } else {
-            hCallback('//dream.tribe.nu/r3/off?q=' + url)
-          }
-        }).catch((error) => {
-          // console.log('saio process error')
-          hCallback('//dream.tribe.nu/r3/off?q=' + url)
-        })
-        return ax
       },
       error (a) {
         setTimeout(function () {
@@ -190,18 +229,26 @@ export default {
         }
       },
       volUp () {
+        let v = 0
         if (store.getters.ytUseVideo && store.getters.isYT) {
-          store.getters.ytObject.setVolume(store.getters.ytObject.getVolume() + 5)
+          v = store.getters.ytObject.getVolume() + 5
+          store.getters.ytObject.setVolume(v)
         } else {
-          DCPlayer.eAudio.volume = Math.min(1, DCPlayer.eAudio.volume + 0.05)
+          v = Math.min(1, DCPlayer.eAudio.volume + 0.05)
+          DCPlayer.eAudio.volume = v
         }
+        return v
       },
       volDown () {
+        let v = 0
         if (store.getters.ytUseVideo && store.getters.isYT) {
-          store.getters.ytObject.setVolume(store.getters.ytObject.getVolume() - 5)
+          v = Math.max(0, store.getters.ytObject.getVolume() - 5)
+          store.getters.ytObject.setVolume(v)
         } else {
-          DCPlayer.eAudio.volume = Math.max(0, DCPlayer.eAudio.volume - 0.05)
+          v = Math.max(0, DCPlayer.eAudio.volume - 0.05)
+          DCPlayer.eAudio.volume = v
         }
+        return v
       }
     }
     DCPlayer.init()
