@@ -1,27 +1,62 @@
 <template>
-  <v-layout row wrap justify-left>
-    <v-flex xs6 lg2 class="ma-0 pa-0">
-      <v-btn icon class="primary" outline>
+  <v-layout 
+    row 
+    wrap 
+    justify-left
+  >
+    <v-flex 
+      xs6 
+      lg2 
+      class="ma-0 pa-0"
+    >
+      <v-btn 
+        icon 
+        class="primary" 
+        outline
+      >
         <v-icon>
           skip_previous
         </v-icon>
       </v-btn>
-      <v-btn icon class="primary" outline>
+      <v-btn 
+        icon 
+        class="primary" 
+        outline
+      >
         <v-icon>
           play_arrow
         </v-icon>
       </v-btn>
-      <v-btn icon class="primary" outline>
+      <v-btn 
+        icon 
+        class="primary" 
+        outline
+      >
         <v-icon>
           skip_next
         </v-icon>
       </v-btn>
     </v-flex>
-    <v-flex xs4 xl9 class="ma-0 pa-0">
-      <v-slider class="pt-2" hide-details color= "primary"></v-slider>
+    <v-flex 
+      xs4 
+      xl9 
+      class="ma-0 pa-0"
+    >
+      <v-slider 
+        class="pt-2" 
+        hide-details 
+        color= "primary"
+      />
     </v-flex>
-    <v-flex xs1 lg10>
-      <v-btn icon class="primary" outline>
+    <v-flex 
+      xs1 
+      lg10
+    >
+      <v-btn 
+        icon 
+        class="primary" 
+        outline
+      >
         <v-icon>
           people
         </v-icon>
