@@ -1,59 +1,76 @@
 <template>
   <div class="?">
-    <v-bottom-nav shift absolute :active.sync="e2" :value="true" color="transparent" v-if="$store.getters.auth_state">
-      <!-- <v-btn :color="$route.name === 'home' ? 'primary' : ''"  value="home" :to="{name: 'searchPage'}">
-        <span>Search</span>
-        <v-icon>search</v-icon>
-      </v-btn> -->
-      <v-btn color="primary" flat value="home" :to="{name: 'historyRecommended', params: {user: UID}}">
+    <v-bottom-nav 
+      :value="true" 
+      absolute 
+      color="transparent"
+    >
+      <v-btn 
+        :to="{name: 'home', params: {user: uid}}" 
+        color="primary" 
+        flat 
+      >
         <span>Home</span>
         <v-icon>home</v-icon>
       </v-btn>
-      <v-btn color="primary" flat value="subsAll" :to="{name: 'subsAll', params: {user: UID}}">
+      <v-btn 
+        :to="{name: 'subsAll', params: {user: uid}}" 
+        color="primary" 
+        flat 
+      >
         <span>Latest</span>
         <v-icon>whatshot</v-icon>
       </v-btn>
-      <v-btn color="primary" flat value="stage" :to="{name: 'stage'}">
+      <v-btn 
+        :to="$store.getters.current_trackID ? {name: 'auto', params: { artist: $store.getters.current_song.artist, trackID: $store.getters.current_song.trackID, source: $store.getters.current_song.source }} : { name: 'auto', params: {source: 'z', artist: 'z', trackID: 'z'} }"
+        flat
+        color="primary" 
+      >
         <span>Current</span>
         <v-icon>music_video</v-icon>
       </v-btn>
-      <v-btn color="primary" flat value="playlistOverview" :to="{name: 'playlistOverview', params: {user: UID}}">
+      <v-btn 
+        :to="{name: 'playlistOverview', params: {user: uid}}" 
+        color="primary" 
+        flat 
+      >
         <span>Playlists</span>
         <v-icon>library_music</v-icon>
       </v-btn>
-      <v-btn color="primary" flat value="userSubOverview" :to="{name: 'userSubOverview', params: {user: UID}}">
+      <v-btn 
+        :to="{name: 'userSubOverview', params: {user: uid}}" 
+        color="primary" 
+        flat 
+      >
         <span>Following</span>
         <v-icon>people</v-icon>
-      </v-btn>
-      <v-btn color="primary" flat value="settings" :to="{name: 'settings'}">
-        <span>Settings</span>
-        <v-icon>settings</v-icon>
       </v-btn>
     </v-bottom-nav>
   </div>
 </template>
 <script>
 /* eslint-disable */
-
-import deleteButton from '@/components/buttons/delete-button'
+import { mapGetters } from 'vuex'
 export default {
-  name: 'x',
-  data () {
-    return {
-      e2: 3
-    }
-  },
-  components: {
-    'delete-button': deleteButton
-  },
-  methods: {
-    x2 () {
-
-    }
-  },
+  name: 'mobFoot',
   computed: {
-    UID () {
-      return this.$DCFB.UID
+    // auto () {
+    //   return this.$store.getters.current_Playlist.length > 0 ? {name: 'auto', params: { artist: this.$store.getters.current_song.artist, trackID: this.$store.getters.current_song.trackID, source: this.$store.getters.current_song.source }} : {name: 'stage'}
+    // },
+    auto () {
+      return this.$store.getters.current_trackID ? {name: 'auto', params: { artist: this.$store.getters.current_song.artist, trackID: this.$store.getters.current_song.trackID, source: this.$store.getters.current_song.source }} : { name: 'stage' }
+    },
+    uid () {
+      return this.$store.getters.uid || Math.random()
+    },
+    active : {
+      get () {
+        // console.log(this.$route)
+        return this.$route.name
+        // return
+      },
+      set (v) {
+      }
     }
   }
 }
