@@ -1,15 +1,25 @@
 <template>
-  <v-flex justify-center align-center class="loading-flex mt-3">
+  <v-flex 
+    justify-center 
+    align-center 
+    class="loading-flex mt-3"
+  >
     <div class="half-circle-spinner">
-      <div class="circle circle-1" :style="circle1Style"></div>
-      <div class="circle circle-2" :style="circle2Style"></div>
+      <div 
+        :style="circle1Style" 
+        class="circle circle-1"
+      />
+      <div 
+        :style="circle2Style" 
+        class="circle circle-2"
+      />
     </div>
   </v-flex>
 </template>
 
 <script>
 export default {
-  name: 'loading',
+  name: 'Loading',
   computed: {
     circle1Style () {
       return Object.assign({
