@@ -1,99 +1,207 @@
 <template>
-  <v-layout row wrap v-show="!show" class="pb-5">
+  <v-layout 
+    v-if="!show" 
+    row 
+    wrap 
+    class="pb-5"
+  >
     <!-- IMAGE -->
-    <v-flex xs12 id="img-bg" v-show="!($store.getters.ytUseVideo && $store.getters.isYT)" @click="$DCPlayer.togglePlay()">
+    <v-flex 
+      v-show="!($store.getters.ytUseVideo && $store.getters.isYT)" 
+      class="nosel"
+      xs12 
+      @click="$DCPlayer.togglePlay()"
+      @contextmenu="$emit('conmen', [$event, [song]])"
+    >
       <div class="pstr-wrapper">
-        <div v-lazy:background-image="song.posterLarge" :style="{ 'background': 'center center no-repeat' }" id="pstr" :key="song.poster"></div>
+        <div 
+          id="pstr" 
+          :style="{ 'background': 'center center no-repeat' , 'background-image' : 'url(' + song.posterLarge + ')'}" 
+          :key="song.poster"
+        />
       </div>
     </v-flex>
-    <v-flex d-flex xs12 v-if="!show">
-      <v-layout row wrap id="dc-padding">
+    <v-flex 
+      dFlex 
+      xs12
+    >
+      <v-layout 
+        id="dc-padding" 
+        row 
+        wrap
+      >
         <!-- SONG TITLE -->
-        <v-flex xs12 class="mt-2">
-          <div class="title fwl text-xs-left">{{$store.getters.current_song.title}}</div >
+        <v-flex 
+          xs12 
+          class="mt-2"
+        >
+          <div class="title fwl text-xs-left"><span>{{ $store.getters.current_song.title }}</span></div >
         </v-flex> 
         <!-- BUTTONS AND UPLOADED DATE/VIEWS AND DIVIDER -->
-        <v-flex xs12 class="stage-btns" :style="stageBorderStyle">
+        <v-flex 
+          :style="stageBorderStyle" 
+          xs12 
+          class="stage-btns"
+        >
           <!-- FLOAT LEFT -->
-          <div class="fl-l blue-grey--text text--lighten-1">
-            {{iViews}}
-            {{$DCAPI.calcDate('', song.uploaded)}}
+          <div class="fl-l blue-grey--text text--lighten-1 mt-3">
+            {{ iViews > -1 ? iViews + ' • ' : '' }}{{ $DCAPI.calcDate('', song.uploaded) }}
           </div>
           <!-- FLOAT RIGHT -->
           <div class="fl-r">
             <!-- LINK -->
-            <v-btn :color="btnCol" @click="($UTILS.copyToClipboard(song.mp32), btnFeedback())" icon>
+            <!-- <v-btn :color="btnCol" @click="($UTILS.copyToClipboard(song.mp32), btnFeedback())" icon>
               <v-icon>link</v-icon>
-            </v-btn>
+            </v-btn> -->
             <!-- YT BUTTON -->
-            <youtube-button></youtube-button>
+            <youtube-button/>
+            
+            <!-- open in a new tab -->
+            <new-tab
+              :song="song"
+            />
             <!-- SHARE BUTTON -->
-            <share-button :song="song" :url="'https://dreamcloud.netlify.com/#/t/' + song.source + '/' + encodeURIComponent(song.artist) + '/' + song.trackID"></share-button>
+            <share-button 
+              :song="song" 
+            />
             <!-- DOWNLOAD BUTTON -->
-            <download-button :links="[song]"></download-button>
+            <download-button :links="[song]"/>
             <!-- ADD TO PLAYLIST -->
-            <add-to-playlist v-if="$store.getters.auth_state" :song="song"></add-to-playlist>
+            <add-to-playlist 
+              v-if="auth_state" 
+              :song="song"
+            />
             <!-- WIDE SCREEN BUTTON -->
-            <v-btn @click="widescreen" icon v-if="$vuetify.breakpoint.lgAndUp">
-              <v-icon>crop_16_9</v-icon>
-            </v-btn>
+            <v-tooltip 
+              v-if="$vuetify.breakpoint.lgAndUp" 
+              top
+            >
+              <v-btn 
+                slot="activator" 
+                icon 
+                @click="widescreen"
+              >
+                <v-icon>{{ bWide ? 'crop_3_2' :'crop_16_9' }}</v-icon>
+              </v-btn>
+              <span>{{ bWide ? 'Wide' :'Narrow' }}</span>
+            </v-tooltip>
             <!-- FULLSCREEN BUTTON -->
-            <v-btn @click="fullscreen" icon>
-              <v-icon>fullscreen</v-icon>
-            </v-btn>
+            <v-tooltip 
+              top
+            >
+              <v-btn 
+                slot="activator" 
+                icon 
+                @click="fullscreen"
+              >
+                <v-icon>fullscreen</v-icon>
+              </v-btn>
+              <span>Fullscreen</span>
+            </v-tooltip>
           </div>
+          
         </v-flex> 
-        <!-- ARTIST PICTURE -->
-        <artist-mini :artistID="song.artistID" :source="song.source" :artist="song.artist" :key="song.artistID"></artist-mini>
         <!-- ARTIST NAME + SONG DESCRIPTION -->
-        <v-flex xs12 lg7 class="title fwl text-xs-left song-meta mt-3 wordbreak">
-          {{ song.artist }}
-          <!-- DESCRIPTION -->
-          <v-flex>
-            <span class="subheading wordbreak fwl preline" v-html="timeToSeconds(_description)"></span>
-          </v-flex>
-          <v-tabs
-            class=""
-            v-model="tab"
-            ref="tabs"
+        <v-flex 
+          xs12 
+          lg8
+          class="text-xs-left song-meta mt-3"
+        >
+          <v-layout 
+            row 
+            wrap
           >
-            <v-tabs-slider color="primary"></v-tabs-slider>
+            <!-- ARTIST PICTURE -->
+            <artist-mini 
+              :artistID="song.artistID" 
+              :source="song.source" 
+              :artist="song.artist" 
+              :key="song.artistID"
+              class="mr-3"
+            />
+            <v-flex xs10>
+              <!-- ARTIST -->
+              <div class="subheading">{{ song.artist }}</div>
+              <!-- DESCRIPTION XSS VULN -->
+              <div 
+                class="subheading fwl wordbreak preline" 
+                v-html="timeToSeconds(description)"
+              />
+            </v-flex>
+          </v-layout>
+          <!-- Tabs -->
+          <v-tabs
+            ref="tabs"
+            v-model="tab"
+            class="mt-3"
+          >
+            <v-tabs-slider color="primary"/>
+            <v-tab>
+              Playlist
+            </v-tab>
             <v-tab >
               Comments
             </v-tab>
             <v-tab>
               Lyrics
             </v-tab>
-            <v-tab v-if="$vuetify.breakpoint.xsOnly">
+            <v-tab 
+              v-if="$vuetify.breakpoint.mdAndDown"
+            >
               Related
             </v-tab>
           </v-tabs>
           
           <v-tabs-items v-model="tab">
             <v-tab-item>
+              <!-- CURRENT PLAYLIST -->
+              <playlist 
+                :songs="$store.getters.current_Playlist"
+                :rows-per-page="-1"
+                @conmen="$emit('conmen', $event)"
+              />
+            </v-tab-item>
+            <v-tab-item>
               <!-- COMMENTS -->
-              <songComments :trackID="song.trackID" :source="song.source"></songComments>
+              <songComments 
+                :trackID="song.trackID" 
+                :source="song.source"
+              />
             </v-tab-item>
             <v-tab-item>
               <!-- LYRICS -->
-              <lyrics :getEm="getLyrics" :title="song.title" :artist="song.artist"></lyrics>
+              <lyrics 
+                :title="song.title" 
+                :artist="song.artist"
+              />
             </v-tab-item>
+            <!-- v-if="$vuetify.breakpoint.mdAndDown" -->
             <v-tab-item v-if="$vuetify.breakpoint.mdAndDown">
               <!-- RELATED -->
-              <related></related>
+              <related 
+                @conmen="$emit('conmen', $event)"
+              />
             </v-tab-item>
           </v-tabs-items>
         </v-flex>
         
         <!-- RELATED -->
-        <related v-if="$vuetify.breakpoint.lgAndUp"></related>
+        <relatedd 
+          v-if="$vuetify.breakpoint.lgAndUp" 
+          @conmen="$emit('conmen', $event)"
+        />
       </v-layout>
     </v-flex>
   </v-layout>
 </template>
+
 <script>
 
-import related from '@/components/stage/meta/related'
+import newTab from '@/components/buttons/open-new-tab'
+
+
+import relatedd from '@/components/stage/meta/related'
+import related from '@/router/related/related'
 import artistMini from '@/components/stage/meta/artist-mini'
 import youtubeVBtn from '@/components/stage/meta/toggle-video-button'
 import songComments from '@/components/stage/meta/comments'
@@ -101,14 +209,25 @@ import lyrics from '@/components/stage/meta/lyrics'
 import addToPlaylist from '@/components/buttons/add-to-playlist.vue'
 import shareButton from '@/components/buttons/share-button'
 import downloadButton from '@/components/buttons/download-button'
-// import explode from '@/components/misc/explode'
+import current from '@/components/stage/meta/current'
+import { mapGetters } from 'vuex'
+
 // /* eslint-disable */
 export default {
-  name: 'dc-stage',
+  name: 'DcStage',
+  watch: {
+    current_trackID: {
+      immediate: true,
+      handler: 'trackChanged'
+    }
+  },
   components: {
     // 'explode': explode,
+    'newTab': newTab,
+    'current': current,
     'songComments': songComments,
     'artist-mini': artistMini,
+    'relatedd': relatedd,
     'related': related,
     'lyrics': lyrics,
     'youtube-button': youtubeVBtn,
@@ -118,7 +237,7 @@ export default {
   },
   data () {
     return {
-      tab: 0,
+      tab: 1,
       bWide: false,
       btnCol: '',
       description: '',
@@ -126,9 +245,12 @@ export default {
     }
   },
   computed: {
-    getLyrics () {
-      return this.tab === 1
-    },
+    ...mapGetters({
+      song: 'current_song',
+      current_trackID: 'current_trackID',
+      ytUseVideo: 'ytUseVideo',
+      auth_state: 'auth_state'
+    }),
     _description: {
       get () {
         return this.description ? this.description : this.song.description
@@ -144,26 +266,21 @@ export default {
     },
     show () {
       return this.ytUseVideo && this.$store.getters.isYT
-    },
-    ytUseVideo () {
-      return this.$store.getters.ytUseVideo
-    },
-    song () {
-      return this.$store.getters.current_song
-    },
-    current_trackID () {
-      return this.$store.getters.current_trackID ? this.$store.getters.current_trackID : 'player'
     }
   },
   methods: {
+    trackChanged () {
+      if (this.$route.name === 'auto') {
+        this.$router.replace({name: 'auto', params: { artist: this.$store.getters.current_song.artist,  trackID: this.$store.getters.current_song.trackID,  source: this.$store.getters.current_song.source }})
+        // return
+      }
+      this.getDesc()
+      this.getPlays()
+    },
     widescreen () {
       this.bWide = !(this.$store.getters.drawLeft || this.$store.getters.drawRight)
-      // if this.bWide() {
       this.$store.commit('drawRight', this.bWide)
       this.$store.commit('drawLeft', this.bWide)
-      // } else {
-      // }
-      // }
     },
     btnFeedback () {
       this.btnCol = 'primary'
@@ -196,7 +313,7 @@ export default {
         return ''
       }
       return (value.replace(/(?:(?:([01]?\d|2[0-3]):)?([0-5]?\d):)([0-5]?\d)/g,
-        `<span class="underline pointer" onClick="document.getElementById('dc-audio').currentTime = '$&'.split(':').reduce((acc,time) => (60 * acc) + +time)">$&</span>`))
+                            `<span class="underline pointer" onClick="document.getElementById('dc-audio').currentTime = '$&'.split(':').reduce((acc,time) => (60 * acc) + +time)">$&</span>`))
     },
     getDesc () {
       if (this.$store.getters.isYT) {
@@ -207,23 +324,12 @@ export default {
         this._description = this.song.description
       }
     }
-  },
-  created () {
-    // console.log('audio created')
-    this.getDesc()
-    this.getPlays()
-  },
-  updated () {
-    this.getDesc()
-    if (this.$store.getters.isYT) {
-      this.getPlays()
-    }
   }
 }
 </script>
 
 <style>
-.pstr-wrapper {position: relative; padding-bottom: 40%; /* 16:9 */  padding-top: 25px;}
+.pstr-wrapper {position: relative; padding-bottom: 38%; /* 16:9 */  padding-top: 25px;}
 .pstr-wrapper #pstr {position: absolute; top: 0; left: 0; width: 100%; height: 100%;}
 
 
@@ -234,25 +340,10 @@ export default {
 
 @media only screen and (min-width: 600px){
   .song-meta {  
-    padding: 0px 20px;
+    padding: 0px 0px 0px 20px;
   }
 }
 
-
-.fl-l{
-  float: left;
-  margin-top: 10px;
-}
-
-#img-bg{
-  background-color:black;
-  /* make div not selectable */
-  -moz-user-select: -moz-none;
-  -khtml-user-select: none;
-  -webkit-user-select: none;
-  -o-user-select: none;
-  user-select: none;
-}
 /* .stage-btns{ */
   /* height: 36px; */
   /* border-bottom: 1px solid teal; */
