@@ -1,57 +1,88 @@
 <template>
   <v-flex class="ma-0 pa-0" >
-    <v-layout row wrap class="ma-0 pa-0">
-      <v-flex xs1 sm1 style="min-width:64px">
-      </v-flex>
-      <v-flex xs11 @click="show = !show" class="pointer body-1">
-        {{show ? `Hide ${totalReplyCount} replies` : `Show ${totalReplyCount} replies`}} 
-        <v-icon size="18">{{show ? 'keyboard_arrow_up' : 'keyboard_arrow_down'}}</v-icon>
-        <orbit v-if="bLoading" class="my-3 ml-4"></orbit>
+    <v-layout 
+      row 
+      wrap 
+      class="ma-0 pa-0"
+    >
+      <v-flex 
+        xs12 
+        class="pointer body-1 comMor" 
+        @click="show = !show"
+      >
+        {{ (show ? 'Hide' : 'Show') + ` ${totalReplyCount} replies` }} 
+        <v-icon size="18">{{ show ? 'keyboard_arrow_up' : 'keyboard_arrow_down' }}</v-icon>
+        <orbit 
+          v-if="bLoading" 
+          class="my-3 ml-4"
+        />
       </v-flex>
     </v-layout>
 
     <v-data-iterator
+      v-if="show"
+      :items="aComments"
       content-tag="v-layout"
       row
       wrap
-      :items="aComments"
       hide-actions
       no-data-text=""
-      v-if="show"
     >
       <v-flex
+        v-if="show"
         slot="item"
         slot-scope="props"
         offset-xs1
         xs11
         class="pr-1 py-1"
-        v-if="show"
       >
-        <v-card class="py-2" flat>
-            <v-layout row>
-              <v-flex xs1 sm1 style="min-width:64px">
-                <v-card-media contain class="pl-2">
-                  <router-link class="body-1 grey--text noDeco" :to="{name: 'artist', params: {source: source, artist: props.item.artist, artistID: props.item.artistID}}">
-                    <v-avatar size="40" color="grey lighten-4">
-                      <img :src="props.item.artistIMG" alt="avatar">
-                    </v-avatar>
+        <v-card 
+          class="py-2" 
+          flat
+        >
+          <v-layout row>
+            <v-flex 
+              xs1 
+              sm1 
+              style="min-width:54px"
+            >
+              <router-link 
+                :to="{name: 'artist', params: {source: source, artist: props.item.artist, artistID: props.item.artistID}}" 
+                class="body-1 grey--text noDeco pl-2"
+              >
+                <v-avatar 
+                  size="40" 
+                  color="grey lighten-4"
+                >
+                  <img 
+                    :src="props.item.artistIMG" 
+                    alt="avatar"
+                  >
+                </v-avatar>
+              </router-link>
+            </v-flex>
+            <v-flex xs11>
+              <div>
+                <div class="body-1 grey--text py-1">
+                  <!-- ARTIST -->
+                  <router-link 
+                    :to="{name: 'artist', params: {source: source, artist: props.item.artist, artistID: props.item.artistID}}" 
+                    class="body-1 grey--text noDeco"
+                  >
+                    <strong>{{ props.item.artist }}</strong> 
                   </router-link>
-                </v-card-media>
-              </v-flex>
-              <v-flex xs11>
-                <div>
-                  <div class="body-1 grey--text">
-                    <router-link class="body-1 grey--text noDeco" :to="{name: 'artist', params: {source: source, artist: props.item.artist, artistID: props.item.artistID}}">
-                      <strong>{{props.item.artist}}</strong> 
-                    </router-link>
-                  <span>{{' ' + $DCAPI.calcDate('', props.item.commentCreated)}}</span></div>
-                  <div class="body-1 preline" v-html="timeToSeconds(props.item.comment)"></div>
-                </div>
-              </v-flex>
-            </v-layout>
+                <!-- CREATED -->
+                <span>{{ ' ' + $DCAPI.calcDate('', props.item.commentCreated) }}</span></div>
+                <div 
+                  class="body-1 preline py-1" 
+                  v-html="timeToSeconds(props.item.comment)"
+                />
+              </div>
+            </v-flex>
+          </v-layout>
           <!-- </v-container> -->
         </v-card>
-        <v-divider v-if="props.index !== aComments.length -1"></v-divider>
+        <v-divider v-if="props.index !== aComments.length -1"/>
       </v-flex>
       <!-- <v-flex slot="footer" v-if="aComments.length && $DCAPI.YTCommentNext">
         <v-btn :loading="bLoading" small block color="transparent" @click="getMore">
@@ -100,7 +131,7 @@ export default {
       if (!value) {
         return ''
       }
-      if (this.$store.getters.ytUseVideo) {
+      if (this.$store.getters.ytUseVideo && this.$store.getters.isYT) {
         return (value.replace(/(?:(?:([01]?\d|2[0-3]):)?([0-5]?\d):)([0-5]?\d)/g,
         `<span class="underline pointer" onClick="window.dcYT.seekTo('$&'.split(':').reduce((acc,time) => (60 * acc) + +time));">$&</span>`))
       } else {
@@ -141,5 +172,8 @@ export default {
 </script>
 
 <style>
-
+.comMor{
+  position: relative;
+  left: 75px !important;
+}
 </style>
