@@ -1,37 +1,54 @@
 <template>
-<v-card>
-  <v-card-title>
-    <div v-if="lyrics" >
+  <v-card>
+    <v-card-title v-if="lyrics">
+      <div class="wordbreak preline">
+        <a 
+          :href="lyricsURL" 
+          class="primary--background"
+        >{{ lyricsURL }}</a>
+        <br>
+        {{ lyrics }}
+      </div>
+    </v-card-title>
+
+    <v-card-title v-else-if="bLoading">
       <!-- <a :href="`${lyricsServer}/search/${this.query}`">{{ `${lyricsServer}/search/${this.query}` }}</a><br> -->
-      <a class="primary--background" :href="lyricsURL">{{ lyricsURL }}</a><br>
-      <div class="wordbreak preline">{{ lyrics }}</div>
-    </div>
-    <div v-else-if="bLoading">
-      <!-- <a :href="`${lyricsServer}/search/${this.query}`">{{ `${lyricsServer}/search/${this.query}` }}</a><br> -->
-      {{loadingText + loadingTextAppend}}
-    </div>
-    <div v-else-if="iTried">
-      <!-- <a :href="`${lyricsServer}/search/${this.ww}`">{{ `${lyricsServer}/search/${this.ww}` }}</a><br> -->
+      {{ loadingText + loadingTextAppend }}
+    </v-card-title>
+
+    <v-card-title v-else-if="iTried">
       No lyrics available
-    </div>
-  </v-card-title>
-</v-card>
+    </v-card-title>
+  
+    <v-card-title 
+      v-else 
+      class="pointer" 
+      @click="(getLyrics)"
+    >
+      Click to load lyrics
+    </v-card-title>
+  </v-card>
 
 </template>
 <script>
 import axios from 'axios'
 // /* eslint-disable */
 export default {
-  name: 'lyrics',
-  props: ['title', 'artist', 'getEm'],
+  name: 'Lyrics',
+  props: {
+    title: {
+      type: String,
+      default: ''
+    },
+    artist: {
+      type: String,
+      default: ''
+    }
+  },
   watch: {
     title: {
       immediate: true,
       handler: 'reset'
-    },
-    getEm: {
-      immediate: true,
-      handler: 'getLyrics'
     }
   },
   data () {
@@ -65,6 +82,7 @@ export default {
       this.bLoading = false
       this.lyrics = ''
       this.iTried = false
+      this.getLyrics()
     },
     animateText () {
       if (this.loadingTextAppend.length === 3) {
@@ -79,7 +97,7 @@ export default {
       clearInterval(intv)
     },
     getLyrics () {
-      if (this.getEm && !this.lyrics) {
+      if (!this.lyrics) {
         var title = this.title
         this.bLoading = true
         this.lyrics = ''
