@@ -1,56 +1,134 @@
 <template>
-  <v-flex xs12 lg4 class="mt-3">
-    <!-- Loading -->
+  <v-flex 
+    xs12 
+    lg4 
+    class="mt-3"
+  >
+    <!-- UP NEXT COMPONENT -->
+    <div v-if="!$store.getters.drawRight">
+      <div class="text-xs-left subheading">
+        Up next
+      </div>
+      <v-flex 
+        xs12
+        @click="($DCPlayer.next(), $DCFB.historyPush($store.getters.current_song))"
+        @contextmenu="$emit('conmen', [$event, upNext])"
+      >
+        <v-card 
+          class="mb-2 pointer" 
+          color=""
+        >
+          <v-layout row>
+            <v-flex xs5>
+              <!-- image -->
+              <v-img
+                :aspect-ratio="upNext.source === 'YouTube' ? 16/9 : 1"
+                :src="upNext.poster"
+                class="fillPlace"
+              >
+                <v-layout
+                  slot="placeholder"
+                  fill-height
+                  align-center
+                  justify-center
+                  ma-0
+                >
+                  <!-- <v-progress-circular 
+                    indeterminate 
+                    color="grey lighten-5"
+                  /> -->
+                </v-layout>
+              </v-img>
+            </v-flex>
+            <v-flex 
+              xs7 
+              class="ma-2 text-xs-left"
+            >
+              <!-- title -->
+              <div class="subheading wordbreak">{{ upNext.title }}</div>
+              <!-- artist -->
+              <div 
+                class="subheading grey--text artist" 
+                @click.stop="$router.push({name: 'artist', params: {source: upNext.source, artist: upNext.artist, artistID: upNext.artistID}})"
+              >{{ upNext.artist }}</div>
+              <!-- duration -->
+              <div class="grey--text">{{ upNext.duration }}</div>
+            </v-flex>
+          </v-layout>
+        </v-card>
+      </v-flex>
+      <v-divider/>
+    </div>
+    <div class="text-xs-left subheading mt-2">
+      Related
+    </div>
     <v-flex 
-      xs2
-      offset-xs5
-      lg2 
-      offset-lg5
       v-if="loading"
+      xs2
+      offset-xs5 
     >
-      <orbit></orbit>
+      <orbit/>
     </v-flex>
     <!-- v-data-iterator -->
     <v-data-iterator
-      hide-actions
-      content-tag='v-layout'
-      row
-      wrap
       :items='items'
       :rows-per-page-items='rowsPerPageItems'
       :pagination.sync='pagination'
       :id='String(song.trackID)'
+      hide-actions
+      content-tag='v-layout'
+      row
+      wrap
     >
       <!-- blank no data -->
-      <v-flex slot="no-data"></v-flex>
-
+      <v-flex slot="no-data"/>
       <!-- related song -->
       <v-flex 
         slot='item'
         slot-scope='props'
         xs12
         @click="play(props.index)"
+        @contextmenu="$emit('conmen', [$event, [props.item]])"
       >
-        <v-card class="mb-2 pointer" color="" >
+        <v-card 
+          class="mb-2 pointer" 
+          color=""
+        >
           <v-layout row>
             <v-flex xs5>
               <!-- image -->
-              <v-card-media
+              <v-img
+                :aspect-ratio="props.item.source === 'YouTube' ? 16/9 : 1"
                 :src="props.item.poster"
-                :height="hai(props.item.source)"
+                class="fillPlace"
               >
-                <!-- <span class="abr15" v-text="props.item.duration"/> -->
-              </v-card-media>
+                <v-layout
+                  slot="placeholder"
+                  fill-height
+                  align-center
+                  justify-center
+                  ma-0
+                >
+                  <!-- <v-progress-circular 
+                    indeterminate 
+                    color="grey lighten-5"
+                  /> -->
+                </v-layout>
+              </v-img>
             </v-flex>
-            <v-flex xs7 class="ml-2">
-              <div>
-                <!-- title -->
-                <div class="subheading text-xs-left wordbreak">{{ props.item.title }}</div>
-                <!-- artist -->
-                <div @click.stop="$router.push({name: 'artist', params: {source: props.item.source, artist: props.item.artist, artistID: props.item.artistID}})" class="text-xs-left grey--text">{{ props.item.artist }}</div>
-                <!-- duration -->
-                <div class="text-xs-left grey--text">{{ props.item.duration }}</div>
-              </div>
+            <v-flex 
+              xs7 
+              class="ma-2 text-xs-left"
+            >
+              <!-- title -->
+              <div class="subheading wordbreak">{{ props.item.title }}</div>
+              <!-- artist -->
+              <div 
+                class="subheading artist" 
+                @click.stop="$router.push({name: 'artist', params: {source: props.item.source, artist: props.item.artist, artistID: props.item.artistID}})"
+              >{{ props.item.artist }}</div>
+              <!-- duration -->
+              <div class="grey--text">{{ props.item.duration }}</div>
             </v-flex>
           </v-layout>
         </v-card>
@@ -65,7 +143,7 @@ import { mapGetters } from 'vuex'
 
 // /* eslint-disable */
 export default {
-  name: 'stage-related',
+  name: 'StageRelated',
   components: {
     'orbit': orbit
   },
@@ -87,6 +165,7 @@ export default {
     ...mapGetters({
       isYT: 'isYT',
       song: 'current_song',
+      upNext: 'next_song',
       trackID: 'current_trackID',
       index: 'index'
     })
@@ -107,11 +186,12 @@ export default {
       this.items = []
       this.$DCAPI.searchInt('', 0, [this.song.source], this.trackID, (d) => {
         this.loading = false
-        if (d[0].trackID === this.trackID) {
-          // alert('removing dupe first')
-          d.shift()
+        if (d.length) {
+          if (d[0].trackID === this.trackID) {
+            d.shift()
+          }
+          this.items = d
         }
-        this.items = d
       }, true, 50)
     },
     hai (source) {
@@ -131,9 +211,9 @@ export default {
   bottom: 5px;
   right:15px;
 }
-.artist{
+/* .artist{
   color: grey;
   float: left;
-}
+} */
 
 </style>
