@@ -1,102 +1,206 @@
 <template>
-  <v-layout row wrap v-show="$store.getters.ytUseVideo && $store.getters.isYT" class="pb-5 ma-0 pa-0">
+  <v-layout 
+    v-show="ytUseVideo && isYT" 
+    row 
+    wrap 
+    class="pb-5 ma-0 pa-0"
+  >
     <!-- VIDEO -->
     <v-flex xs12>
       <div class="video-wrapper">
-        <div id="player"></div>
+        <div id="player"/>
       </div>
     </v-flex>
-    <v-flex d-flex xs12 v-if="$store.getters.ytUseVideo && $store.getters.isYT">
-      <v-layout row wrap id="dc-padding">
+    <v-flex 
+      v-if="ytUseVideo && isYT" 
+      dFlex 
+      xs12
+    >
+      <v-layout 
+        id="dc-padding" 
+        row 
+        wrap
+      >
         <!-- SONG TITLE -->
-        <v-flex xs12 class="mt-2">
-          <div class="title fwl text-xs-left">{{$store.getters.current_song.title}}</div >
+        <v-flex 
+          xs12 
+          class="mt-2"
+        >
+          <div class="title fwl text-xs-left">{{ song.title }}</div >
         </v-flex>
         <!-- BUTTONS AND UPLOADED DATE/VIEWS AND DIVIDER -->
-        <v-flex xs12 class="stage-btns" :style="stageBorderStyle">
-          <div class="fl-l blue-grey--text text--lighten-1">
-            {{iViews}}
-            {{$DCAPI.calcDate('', song.uploaded)}}
+        <v-flex 
+          :style="stageBorderStyle" 
+          xs12 
+          class="stage-btns"
+        >
+          <!-- FLOAT LEFT -->
+          <div class="fl-l blue-grey--text text--lighten-1 mt-3">
+            {{ iViews }} • {{ $DCAPI.calcDate('', song.uploaded) }}
           </div>
+          <!-- FLOAT RIGHT -->
           <div class="fl-r">
             <!-- LINK -->
-            <v-btn :color="btnCol" @click="($UTILS.copyToClipboard(song.mp32), btnFeedback())" icon>
+            <!-- <v-btn :color="btnCol" @click="($UTILS.copyToClipboard(song.mp32), btnFeedback())" icon>
               <v-icon>link</v-icon>
-            </v-btn>
-            <!-- CLOSED CAPTIONS -->
-            <v-btn icon @click="toggleCC">
-              <v-icon :color="cc ? 'white' : 'grey'">subtitles</v-icon>
-            </v-btn>
+            </v-btn> -->
+
+
             <!-- YT BUTTON -->
-            <youtube-button></youtube-button>
+            <youtube-button slot="activator" />
+
+            <!-- Open in a new tab -->
+            <new-tab
+              :song="song"
+            />
+            <!-- CLOSED CAPTIONS -->
+            <v-tooltip 
+              v-if="$vuetify.breakpoint.lgAndUp" 
+              top
+            >
+              <v-btn 
+                slot="activator" 
+                icon 
+                @click="toggleCC"
+              >
+                <v-icon :color="cc ? 'white' : 'grey'">subtitles</v-icon>
+              </v-btn>
+              <span>Subtitles</span>
+            </v-tooltip>
             <!-- SHARE BUTTON -->
-            <share-button :song="song" :url="'https://dreamcloud.netlify.com/#/t/' + song.source + '/' + encodeURIComponent(song.artist) + '/' + song.trackID"></share-button>
+            <share-button 
+              :song="song" 
+            />
             <!-- DOWNLOAD BUTTON -->
-            <download-button :links="[song]"></download-button>
+            <download-button :links="[song]"/>
             <!-- ADD TO PLAYLIST -->
-            <add-to-playlist v-if="$store.getters.auth_state" :song="song"></add-to-playlist>
+            <add-to-playlist 
+              v-if="$store.getters.auth_state" 
+              :song="song"
+            />
             <!-- WIDE SCREEN BUTTON -->
-            <v-btn @click="widescreen" icon v-if="$vuetify.breakpoint.lgAndUp">
-              <v-icon>{{bWide ? 'crop_3_2' :'crop_16_9'}}</v-icon>
-            </v-btn>
+            <v-tooltip 
+              v-if="$vuetify.breakpoint.lgAndUp" 
+              top
+            >
+              <v-btn 
+                slot="activator" 
+                icon 
+                @click="widescreen"
+              >
+                <v-icon>{{ bWide ? 'crop_3_2' :'crop_16_9' }}</v-icon>
+              </v-btn>
+              <span>{{ bWide ? 'Wide' :'Narrow' }}</span>
+            </v-tooltip>
             <!-- FULLSCREEN BUTTON -->
-            <v-btn @click="fullscreen" icon>
-              <v-icon>fullscreen</v-icon>
-            </v-btn>
+            <v-tooltip 
+              top
+            >
+              <v-btn 
+                slot="activator" 
+                icon 
+                @click="fullscreen"
+              >
+                <v-icon>fullscreen</v-icon>
+              </v-btn>
+              <span>Fullscreen</span>
+            </v-tooltip>
           </div>
         </v-flex>
-        <!-- ARTIST PICTURE -->
-        <artist-mini :artistID="song.artistID" :source="song.source" :artist="song.artist" :key="song.artistID"></artist-mini>
-        <!-- ARTIST NAME + SONG DESCRIPTION -->
-        <v-flex xs12 lg7 class="title fwl text-xs-left song-meta mt-3">
-          {{ song.artist }}
-          <!-- DESCRIPTION -->
-          <v-flex xs12>
-            <span class="subheading fwl wordbreak preline" v-html="timeToSeconds(description)"></span>
-          </v-flex>
-
-          <v-tabs
-            class=""
-            v-model="tab"
-            ref="tabs"
+        <v-flex 
+          xs12 
+          lg8
+          class="text-xs-left mt-3 song-meta"
+        >
+          <v-layout 
+            row 
+            wrap
           >
-            <v-tabs-slider color="primary"></v-tabs-slider>
+            <!-- <div class="text-xs-left song-meta mt-3 wordbreak"> -->
+            <!-- </div> -->
+            <!-- DESCRIPTION -->
+            <!-- ARTIST PICTURE -->
+            <artist-mini 
+              :artistID="song.artistID" 
+              :source="song.source" 
+              :artist="song.artist" 
+              :key="song.artistID"
+              class="mr-3"
+            />
+            <v-flex xs10>
+              <div class="subheading">{{ song.artist }}</div>
+              <div 
+                class="subheading fwl wordbreak preline" 
+                v-html="timeToSeconds(description)"
+              />
+            </v-flex>
+          </v-layout>
+          <v-tabs
+            ref="tabs"
+            v-model="tab"
+            class="mt-3"
+          >
+            <v-tabs-slider color="primary"/>
+            <v-tab >
+              Playlist
+            </v-tab>
             <v-tab >
               Comments
             </v-tab>
             <v-tab>
               Lyrics
             </v-tab>
-            <v-tab v-if="$vuetify.breakpoint.xsOnly">
+            <v-tab v-if="$vuetify.breakpoint.mdAndDown">
               Related
             </v-tab>
           </v-tabs>
           
           <v-tabs-items v-model="tab">
             <v-tab-item>
+              <!-- CURRENT PLAYLIST -->
+              <playlist 
+                :songs="$store.getters.current_Playlist" 
+                :rows-per-page="-1"
+                @conmen="$emit('conmen', $event)"
+              />
+            </v-tab-item>
+            <v-tab-item>
               <!-- COMMENTS -->
-              <songComments :trackID="song.trackID" :source="song.source"></songComments>
+              <songComments 
+                :trackID="song.trackID" 
+                :source="song.source"
+              />
             </v-tab-item>
             <v-tab-item>
               <!-- LYRICS -->
-              <lyrics :getEm="getLyrics" :title="song.title" :artist="song.artist"></lyrics>
+              <lyrics 
+                :title="song.title" 
+                :artist="song.artist"
+              />
             </v-tab-item>
             <v-tab-item v-if="$vuetify.breakpoint.mdAndDown">
               <!-- RELATED -->
-              <related></related>
+              <related 
+                @conmen="$emit('conmen', $event)"
+              />
             </v-tab-item>
           </v-tabs-items>
         </v-flex>
         
         <!-- RELATED -->
-        <related v-if="$vuetify.breakpoint.lgAndUp"></related>
+        <relatedd 
+          v-if="$vuetify.breakpoint.lgAndUp" 
+          @conmen="$emit('conmen', $event)"
+        />
       </v-layout>
     </v-flex>
   </v-layout>
 </template>
 <script>
+import newTab from '@/components/buttons/open-new-tab'
 
-import related from '@/components/stage/meta/related'
+import related from '@/router/related/related'
+import relatedd from '@/components/stage/meta/related'
 import artistMini from '@/components/stage/meta/artist-mini'
 import youtubeVBtn from '@/components/stage/meta/toggle-video-button'
 import songComments from '@/components/stage/meta/comments'
@@ -104,22 +208,31 @@ import lyrics from '@/components/stage/meta/lyrics'
 import addToPlaylist from '@/components/buttons/add-to-playlist.vue'
 import shareButton from '@/components/buttons/share-button'
 import downloadButton from '@/components/buttons/download-button'
-
+import current from '@/components/stage/meta/current'
 import { mapGetters } from 'vuex'
 
 /* eslint-disable */
 export default {
   name: 'video-stage',
-  created () {
+  beforeCreate () {
     var tag = document.createElement('script')
     tag.src = 'https://www.youtube.com/iframe_api'
     var fst = document.getElementsByTagName('script')[0]
     fst.parentNode.insertBefore(tag, fst)
   },
+  watch: {
+  trackID: {
+      immediate: true,
+      handler: 'trackChanged'
+    }
+  },
   components: {
+    'newTab': newTab,
+    'current': current,
     'artist-mini': artistMini,
     'youtube-button': youtubeVBtn,
     'related': related,
+    'relatedd': relatedd,
     'lyrics': lyrics,
     'songComments': songComments,
     'add-to-playlist': addToPlaylist,
@@ -128,33 +241,26 @@ export default {
   },
   computed: {
     ...mapGetters({
-      current_song: 'current_song',
+      song: 'current_song',
       index: 'index',
       hash: 'hash',
-      current_song: 'current_song'
+      trackID: 'current_trackID',
+      ytUseVideo: 'ytUseVideo',
+      ytObject: 'ytObject',
+      drawLeft: 'drawLeft',
+      showVideo: 'showVideo',
+      drawRight: 'drawRight',
+      isYT: 'isYT'
     }),
-    getLyrics () {
-      return this.tab === 1
-    },
     stageBorderStyle () {
       return {
         'border-bottom': '1px solid ' + this.$vuetify.theme.primary,
       }
-    },
-    song () {
-      return this.$store.getters.current_song
-    },
-    showClass () {
-      return { 'hidden': this.$store.getters.ytShowVideo }
-      // return this.$store.getters.ytShowVideo
-    },
-    current_trackID () {
-      return this.$store.getters.current_trackID ? this.$store.getters.current_trackID : 'player'
     }
   },
   data () {
     return {
-      tab: 0,
+      tab: 1,
       bWide: false,
       btnCol: '',
       iViews: '', 
@@ -170,7 +276,7 @@ export default {
   },
   methods: {
     widescreen () {
-      this.bWide = !(this.$store.getters.drawLeft || this.$store.getters.drawRight)
+      this.bWide = !(this.drawLeft || this.drawRight)
       this.$store.commit('drawRight', this.bWide)
       this.$store.commit('drawLeft', this.bWide)
     },
@@ -181,7 +287,7 @@ export default {
       }, 2000)
     },
     getPlays () {
-      this.$DCAPI.getSongPlays(this.song.trackID, this.song.source, (data) => {
+      this.$DCAPI.getSongPlays(this.trackID, this.song.source, (data) => {
         this.iViews = this.makeFriendly(data)
       })
     },
@@ -216,7 +322,7 @@ export default {
       this.$UTILS.toggleFullscreen('player')
     },
     getDesc () {
-      this.$DCAPI.getSongDescription(this.current_trackID, this.song.source, (resp) => {
+      this.$DCAPI.getSongDescription(this.trackID, this.song.source, (resp) => {
         this.description = resp.items[0].snippet.description.trim()
       })
     },
@@ -226,7 +332,7 @@ export default {
         setTimeout(this.ytBind, 100)
         return
       }
-      if (!this.$store.getters.ytUseVideo) {
+      if (!this.ytUseVideo) {
         return
       }
       
@@ -234,8 +340,7 @@ export default {
       if (!this.yt) {
         this.yt = new YT.Player('player', {
           width: '100%',
-          videoId: this.current_trackID,
-          
+          videoId: this.trackID,
           enablejsapi: 1,
           playerVars: {
             autoplay: 1,        // Auto-play the video on load
@@ -264,7 +369,7 @@ export default {
     ytReady (state) {
       this.$store.commit('ytObject', state.target)
       // this.$store.commit('ytState', state.data)
-      this.$store.getters.ytObject.playVideo()
+      this.ytObject.playVideo()
       window.dcYT = this.yt
     },
     ytChanged (state) {
@@ -272,48 +377,47 @@ export default {
       this.$store.commit('ytState', state.data)
       // if playing set duration amd interval to set current time.
       if (state.data === 1) {
-        this.$store.commit('ytDuration', this.$store.getters.ytObject.getDuration())
+        this.$store.commit('ytDuration', this.ytObject.getDuration())
         this.interval = setInterval(() => {
-          this.$store.commit('ytCurrentTime', this.$store.getters.ytObject.getCurrentTime())
-        }, 250)
+          this.$store.commit('ytCurrentTime', this.ytObject.getCurrentTime())
+        }, 500)
+      // ENDED
       } else if (state.data === 0) {
         clearInterval(this.interval)
         this.$store.commit('ytCurrentTime', 0)
         this.$DCPlayer.next()
-        this.$DCFB.historyPush(this.$store.getters.current_song)
+        this.$DCFB.historyPush(this.song)
       } else { // if (state.data === 5 || state.data === 3 || state.data === 2) {
         clearInterval(this.interval)
       }
-    }
-  },
-  mounted () {
-    if (this.$store.getters.isYT) {
-      // console.log('vid mounted')
-      this.ytBind()
-      this.getDesc()
-      this.getPlays()
-    }
-  },
-  updated () {
-    // if new song
-    if (this.$store.getters.isYT && this.currentID !== this.current_trackID && this.$store.getters.ytUseVideo && !this.$store.getters.ytSwitchTime) {
-      // console.log('vid updated')
-      this.currentID = this.current_trackID
-      // if not already attached to iframe
-      if (!this.$store.getters.ytObject.hasOwnProperty('loadVideoById')) {
-        this.ytBind()
-      } else {
-        this.$store.getters.ytObject.loadVideoById(this.current_trackID)
+    },
+    trackChanged () {
+      if (this.isYT && this.ytUseVideo) {
+        if (this.$route.name === 'auto') {
+          // this.$router.push({name: 'stage'})
+          this.$router.replace({name: 'auto', params: { artist: this.$store.getters.current_song.artist,  trackID: this.$store.getters.current_song.trackID,  source: this.$store.getters.current_song.source }})
+        }
+
+        // console.log('changing song')
+        if (!this.ytObject.hasOwnProperty('loadVideoById')) {
+          this.ytBind()
+        } else {
+          this.ytObject.loadVideoById(this.trackID)
+        }
+        this.getPlays()
+        this.getDesc()
+        this.$DCPlayer.pause()
       }
-      this.getPlays()
-      this.getDesc()
-      this.$DCPlayer.pause()
-    } else {
-      // This stores the last trackID to ensure we don't load the same video twice by accident.
-      // It's tied to the div above, there's a much better way to do this.
-      this.currentID = this.current_trackID
     }
-  }
+  },
+  // mounted () {
+  //   if (this.$store.getters.isYT) {
+  //     // console.log('vid mounted')
+  //     this.ytBind()
+  //     this.getDesc()
+  //     this.getPlays()
+  //   }
+  // }
 }
 </script>
 
@@ -324,18 +428,13 @@ export default {
   }
 }
 
-.video-wrapper {position: relative; padding-bottom: 40%; /* 56.25% 16:9 */  padding-top: 25px;}
+.video-wrapper {position: relative; padding-bottom: 38%; /* 56.25% 16:9 */  padding-top: 25px;}
 .video-wrapper iframe {position: absolute; top: 0; left: 0; width: 100%; height: 100%;}
 
 
 .video-wrapper {
   background-color: black;
   /* height: 90%; */
-}
-
-.fl-l{
-  float: left;
-  margin-top: 10px;
 }
 
 #img-bg{
