@@ -1,7 +1,8 @@
 export default {
   state: {
-    ytUseVideo: true,
-    ytSwitchTime: false,
+    dcIsPlaying: false,
+    dcIsLoading: false,
+    ytUseVideo: false,
     ytObject: [],
     ytDuration: 0,
     ytCurrentTime: 0,
@@ -16,21 +17,49 @@ export default {
     }
   },
   mutations: {
-    ytSwitchTime: (state, payload) => { state.ytSwitchTime = payload },
-    ytUseVideo: (state, payload) => { state.ytUseVideo = payload },
-    ytObject: (state, payload) => { state.ytObject = payload },
-    ytState: (state, payload) => { state.ytState = payload },
-    ytDuration: (state, payload) => { state.ytDuration = payload },
-    ytCurrentTime: (state, payload) => { state.ytCurrentTime = payload },
+    dcIsLoading(state, payload) {
+      state.dcIsLoading = payload
+    },
+    dcIsPlaying(state, payload) {
+      state.dcIsPlaying = payload
+    },
+    ytUseVideo: (state, payload) => {
+      state.ytUseVideo = payload
+    },
+    ytObject: (state, payload) => {
+      state.ytObject = payload
+    },
+    ytState: (state, payload) => {
+      state.ytState = payload
+    },
+    ytDuration: (state, payload) => {
+      state.ytDuration = payload
+    },
+    ytCurrentTime: (state, payload) => {
+      state.ytCurrentTime = payload
+    },
     ytStopVideo: state => {
       if (state.ytState === 1) {
-        try { state.ytObject.stopVideo() } catch (err) {}
+        if (typeof state.ytObject.stopVideo === 'function') {
+          state.ytObject.stopVideo()
+        }
       }
     },
-    ytToggleVideo: (state, payload) => { state.ytUseVideo = state.ytUseVideo = !state.ytUseVideo }
+    ytToggleVideo: (state) => {
+      state.ytUseVideo = state.ytUseVideo = !state.ytUseVideo
+    }
   },
 
   getters: {
+    isLoading: (state, getters) => state.ytUseVideo && getters.isYT ? getters.ytIsLoading : state.dcIsLoading,
+    isPlaying: (state, getters) => state.ytUseVideo && getters.isYT ? getters.ytIsPlaying : state.dcIsPlaying,
+    // if (state.ytUseVideo) {
+    //   return getters.ytIsLoading
+    // } else {
+    //   return state.dcIsLoading
+    // }
+    dcIsLoading: state => state.dcIsLoading,
+    dcIsPlaying: state => state.dcIsPlaying,
     ytUseVideo: state => state.ytUseVideo,
     ytObject: state => state.ytObject,
     ytState: state => state.ytState,
@@ -38,6 +67,8 @@ export default {
     ytCurrentTime: state => state.ytCurrentTime,
     ytIsPlaying: state => state.ytState === 1,
     ytIsPaused: state => state.ytState === 2,
+    ytIsLoading: (state) => state.ytState === 3,
+    // ytIsLoading: (state, getters) => !getters.ytIsPlaying && !getters.ytIsPaused,
     isYT: (state, getters) => getters.current_source === 'YouTube'
     // ytDuration: state => state.ytObject.getDuration(),
     // ytCurrentTime: state => state.ytObject.getCurrentTime()
